fixup! EZP-26300: Add destination content href to Relation(List) REST field value
