+++ eZ/Publish/Core/REST/Common/FieldTypeProcessor/BaseRelationProcessor.php
@@ -5,7 +5,30 @@
  */
 namespace eZ\Publish\Core\REST\Common\FieldTypeProcessor;
 
-class BaseRelationProcessor
-{
+use Symfony\Component\Routing\RouterInterface;
 
+abstract class BaseRelationProcessor
+{
+    /**
+     * @var \Symfony\Component\Routing\RouterInterface
+     */
+    private $router;
+
+    public function setRouter(RouterInterface $router)
+    {
+        $this->router = $router;
+    }
+
+    public function canMapContentHref()
+    {
+        return isset($this->router);
+    }
+
+    public function mapToContentHref($contentId)
+    {
+        return $this->router->generate(
+            'ezpublish_rest_loadContent',
+            ['contentId' => $contentId]
+        );
+    }
 }
