+++ eZ/Publish/Core/REST/Common/FieldTypeProcessor/RelationListProcessor.php
@@ -14,18 +14,8 @@ use eZ\Publish\Core\REST\Common\FieldTypeProcessor;
 use eZ\Publish\Core\FieldType\RelationList\Type;
 use Symfony\Component\Routing\RouterInterface;
 
-class RelationListProcessor extends FieldTypeProcessor
+class RelationListProcessor extends BaseRelationProcessor
 {
-    /**
-     * @var \Symfony\Component\Routing\RouterInterface
-     */
-    private $router;
-
-    public function setRouter(RouterInterface $router)
-    {
-        $this->router = $router;
-    }
-
     /**
      * In addition to the list of destinationContentIds, adds a destinationContentHrefs
      * array, with matching content uris.
@@ -36,21 +26,20 @@ class RelationListProcessor extends FieldTypeProcessor
      */
     public function postProcessValueHash($outgoingValueHash)
     {
-        if (!isset($outgoingValueHash['destinationContentIds']) || !is_array($outgoingValueHash['destinationContentIds'])) {
+        if (
+            !isset($outgoingValueHash['destinationContentIds']) ||
+            !is_array($outgoingValueHash['destinationContentIds']) ||
+            !$this->canMapContentHref()
+        ) {
             return $outgoingValueHash;
         }
 
-        if (isset($this->router)) {
-            $outgoingValueHash['destinationContentHrefs'] = array_map(
-                function ($contentId) {
-                    return $this->router->generate(
-                        'ezpublish_rest_loadContent',
-                        ['contentId' => $contentId]
-                    );
-                },
-                $outgoingValueHash['destinationContentIds']
-            );
-        }
+        $outgoingValueHash['destinationContentHrefs'] = array_map(
+            function ($contentId) {
+                return $this->mapToContentHref($contentId);
+            },
+            $outgoingValueHash['destinationContentIds']
+        );
 
         return $outgoingValueHash;
     }
